requirements.txt
# Core dependencies for the stock data pipeline
requests>=2.31.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0

# Data processing and validation
pandas>=2.1.0
numpy>=1.24.0

# Security
cryptography>=41.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.1.0
black>=23.7.0
